群聊

+++ IMiPhone/DPChatMessage.h
@@ -11,11 +11,11 @@
 @interface DPChatMessage : NSObject 
 
 @property (nonatomic, retain) NSString *stage;
-@property (nonatomic) long targetId;
+@property (nonatomic) long long targetId;
 @property (nonatomic) NSInteger msgType;
 @property (nonatomic, retain) NSString *content;
-@property (nonatomic) long mid;
-@property (nonatomic) long senderUid;
+@property (nonatomic) long long mid;
+@property (nonatomic) long long senderUid;
 @property (nonatomic, retain) NSString *sendTime;
